+++ idf_component_manager/version_solver/version_solver.py
@@ -1,6 +1,8 @@
 # SPDX-FileCopyrightText: 2022-2023 Espressif Systems (Shanghai) CO LTD
 # SPDX-License-Identifier: Apache-2.0
 
+import os
+
 from idf_component_tools.errors import DependencySolveError, SolverError
 from idf_component_tools.manifest import (
     ComponentRequirement,
@@ -11,13 +13,13 @@ from idf_component_tools.manifest import (
 )
 from idf_component_tools.sources import LocalSource
 
-from ..utils import print_info
+from ..utils import print_info, print_warn
 from .helper import PackageSource
 from .mixology.package import Package
 from .mixology.version_solver import VersionSolver as Solver
 
 try:
-    from typing import Callable, Optional
+    from typing import Callable
 except ImportError:
     pass
 
@@ -29,7 +31,7 @@ class VersionSolver(object):
     """
 
     def __init__(self, requirements, old_solution=None, component_solved_callback=None):
-        # type:(ProjectRequirements, Optional[SolvedManifest], Optional[Callable[[], None]]) -> None
+        # type:(ProjectRequirements, SolvedManifest | None, Callable[[], None] | None) -> None
         self.requirements = requirements
         self.old_solution = old_solution
         self.component_solved_callback = component_solved_callback
@@ -45,10 +47,14 @@ class VersionSolver(object):
         # scan all root local requirements
         # root local requirements defined in the file system manifest files
         # would have higher priorities
+
+        # scan all root local requirements
         for manifest in self.requirements.manifests:
             for requirement in manifest.dependencies:  # type: ComponentRequirement
                 if isinstance(requirement.source, LocalSource):
-                    _recorded_requirement = self._local_root_requirements.get(requirement.name)
+                    _recorded_requirement = self._local_root_requirements.get(
+                        requirement.build_name
+                    )
                     if _recorded_requirement:
                         # can't specify two different root local source dependencies
                         if _recorded_requirement.source.hash_key != requirement.source.hash_key:
@@ -58,7 +64,23 @@ class VersionSolver(object):
                                 )
                             )
                     else:
-                        self._local_root_requirements[requirement.name] = requirement
+                        self._local_root_requirements[requirement.build_name] = requirement
+
+        # scan all root local components
+        for manifest in self.requirements.manifests[1:]:
+            # add itself as highest priority component
+            if manifest.name and manifest.version and manifest._manifest_manager:
+                _source = LocalSource({'path': os.path.dirname(manifest._manifest_manager.path)})
+                self._source.add(
+                    Package(manifest.name, _source),
+                    str(manifest.version),
+                    deps=self._component_dependencies_with_local_precedence(
+                        manifest.dependencies, manifest.name
+                    ),
+                )
+                self._local_root_requirements[manifest.name] = ComponentRequirement(
+                    manifest.name, _source, str(manifest.version)
+                )
 
         for manifest in self.requirements.manifests:
             self.solve_manifest(manifest)
@@ -80,27 +102,12 @@ class VersionSolver(object):
         )
 
     def solve_manifest(self, manifest):  # type: (Manifest) -> None
-        for requirement in manifest.dependencies:  # type: ComponentRequirement
-            # replace root requirement to local one if exists
-            if (
-                not isinstance(requirement.source, LocalSource)
-                and requirement.name in self._local_root_requirements
-            ):
-                _requirement = self._local_root_requirements[requirement.name]
-                print_info(
-                    'replace manifest dependency {} to {}'.format(
-                        requirement,
-                        self._local_root_requirements[requirement.name],
-                    )
-                )
-            else:
-                _requirement = requirement
-
-            self._source.root_dep(
-                Package(_requirement.name, _requirement.source), _requirement.version_spec
-            )
+        for dep in self._dependencies_with_local_precedence(
+            manifest.dependencies, manifest_path=manifest.path
+        ):
+            self._source.root_dep(Package(dep.name, dep.source), dep.version_spec)
             try:
-                self.solve_component(_requirement)
+                self.solve_component(dep, manifest_path=manifest.path)
             except DependencySolveError as e:
                 raise SolverError(
                     'Solver failed processing dependency "{dependency}" '
@@ -114,7 +121,9 @@ class VersionSolver(object):
                     '\n{original_error}'.format(path=manifest.path, original_error=str(e))
                 )
 
-    def solve_component(self, requirement):  # type: (ComponentRequirement) -> None
+    def solve_component(
+        self, requirement, manifest_path=None
+    ):  # type: (ComponentRequirement, str | None) -> None
         if requirement in self._solved_requirements:
             return
 
@@ -125,29 +134,16 @@ class VersionSolver(object):
                 target=self.requirements.target,
             )
         except Exception as e:
-            raise DependencySolveError(str(e), dependency=requirement.name)
+            print_warn(e)
+            return
 
         for version in cmp_with_versions.versions:
             if requirement.source.is_overrider:
-                self._overriders.add(requirement.name)
+                self._overriders.add(requirement.build_name)
 
-            deps = {}
-            for req in version.dependencies:
-                # replace version requirement to local one if exists
-                if req.name in self._local_root_requirements:
-                    print_info(
-                        'replace component {}({}) dependency {} to {}'.format(
-                            requirement.name,
-                            version.text,
-                            req,
-                            self._local_root_requirements[req.name],
-                        )
-                    )
-                    _req = self._local_root_requirements[req.name]
-                else:
-                    _req = req
-
-                deps[Package(_req.name, _req.source)] = _req.version_spec
+            deps = self._component_dependencies_with_local_precedence(
+                version.dependencies, component_name=requirement.name, manifest_path=manifest_path
+            )
 
             self._source.add(
                 Package(requirement.name, requirement.source),
@@ -163,3 +159,47 @@ class VersionSolver(object):
 
         if self.component_solved_callback:
             self.component_solved_callback()
+
+    def _dependencies_with_local_precedence(
+        self,
+        dependencies,  # type: list[ComponentRequirement]
+        component_name=None,  # type: str | None
+        manifest_path=None,  # type: str | None
+    ):  # type: (...) -> list[ComponentRequirement]
+        deps = []  # type: list[ComponentRequirement]
+        for dep in dependencies:
+            # replace version dependencies to local one if exists
+            # use build_name in both recording and replacing
+            if dep.build_name in self._local_root_requirements:
+                print_info(
+                    'Using component placed at {} for dependency {}{}{}'.format(
+                        # must be a local source here
+                        self._local_root_requirements[dep.build_name].source._path,  # type: ignore
+                        dep,
+                        '(introduced by component {})'.format(component_name)
+                        if component_name
+                        else '',
+                        ', specified in {}'.format(manifest_path) if manifest_path else '',
+                    )
+                )
+                deps.append(self._local_root_requirements[dep.build_name])
+            else:
+                deps.append(dep)
+
+        return deps
+
+    def _component_dependencies_with_local_precedence(
+        self,
+        dependencies,  # type: list[ComponentRequirement]
+        component_name=None,  # type: str | None
+        manifest_path=None,  # type: str | None
+    ):  # type: (...) -> dict[Package, str]
+        deps = {}  # type: dict[Package, str]
+        for dep in self._dependencies_with_local_precedence(
+            dependencies,
+            component_name,
+            manifest_path,
+        ):
+            deps[Package(dep.name, dep.source)] = dep.version_spec
+
+        return deps